@@ -53,9 +53,9 @@ class TorchAggregator(Aggregator):
     Basic aggregator using built-in torch methods to perform aggregation
     """
 
-    _method: Callable
+    _method: Callable[[torch.Tensor, int], torch.Tensor]
 
-    def aggregate(self, results: torch.Tensor, mask:torch.Tensor) -> torch.Tensor:
+    def aggregate(self, results: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
         # TODO: check dimension with explicit jax typing for results parameter
         return self._method(results, dim=0)
 
@@ -64,6 +64,7 @@ class MeanAggregator(TorchAggregator):
     """
     Mean of attributions
     """
+
     _method = torch.mean
 
 
@@ -71,8 +72,9 @@ class SquaredMeanAggregator(Aggregator):
     """
     Mean of squares of attributions
     """
+
     @staticmethod
-    def _method(results: torch.Tensor, dim:int=0) -> torch.Tensor:
+    def _method(results: torch.Tensor, dim: int = 0) -> torch.Tensor:
         return torch.mean(torch.square(results), dim=dim)
 
 
@@ -80,6 +82,7 @@ class SumAggregator(TorchAggregator):
     """
     Sum of attributions
     """
+
     _method = torch.sum
 
 
@@ -87,6 +90,7 @@ class VarianceAggregator(TorchAggregator):
     """
     Variance of attributions
     """
+
     _method = torch.var
 
 
@@ -94,13 +98,16 @@ class MaskwiseMeanAggregator(Aggregator):
     """
     TODO : add docstring
     """
+
     def aggregate(self, results: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
         return torch.einsum("pt,pl->tl", results, mask) / mask.sum(dim=1)
+
 
 class OcclusionAggregator(Aggregator):
     """
     TODO : add docstring
     """
+
     def aggregate(self, results: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
         # first prediction is reference, unmodified input
         scores = results[..., 0, :] - results[..., 1:, :]
